# openapi.yaml
openapi: 3.1.0
info:
  title: Hypeman API
  description: Generic API for managing VM lifecycle using Cloud Hypervisor with OCI-based workloads
  version: 0.2.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: Lower-level error code providing more specific detail
          example: invalid_input
        message:
          type: string
          description: Further detail about the error
          example: Image not found
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Application-specific error code (machine-readable)
          example: bad_request
        message:
          type: string
          description: Human-readable error description for debugging
          example: "Missing required field: image"
        details:
          type: array
          description: Additional error details (for multiple errors)
          items:
            $ref: "#/components/schemas/ErrorDetail"
        inner_error:
          $ref: "#/components/schemas/ErrorDetail"
    InstanceState:
      type: string
      enum: [Created, Running, Paused, Shutdown, Stopped, Standby, Unknown]
      description: |
        Instance state:
        - Created: VMM created but not started (Cloud Hypervisor native)
        - Running: VM is actively running (Cloud Hypervisor native)
        - Paused: VM is paused (Cloud Hypervisor native)
        - Shutdown: VM shut down but VMM exists (Cloud Hypervisor native)
        - Stopped: No VMM running, no snapshot exists
        - Standby: No VMM running, snapshot exists (can be restored)
        - Unknown: Failed to determine state (see state_error for details)
    VolumeMount:
      type: object
      required: [volume_id, mount_path]
      properties:
        volume_id:
          type: string
          description: Volume identifier
          example: vol-abc123
        mount_path:
          type: string
          description: Path where volume is mounted in the guest
          example: /mnt/data
        readonly:
          type: boolean
          description: Whether volume is mounted read-only
          default: false
        overlay:
          type: boolean
          description: Create per-instance overlay for writes (requires readonly=true)
          default: false
        overlay_size:
          type: string
          description: Max overlay size as human-readable string (e.g., "1GB"). Required if overlay=true.
          example: "1GB"
...
    CreateInstanceRequest:
      type: object
      required: [name, image]
      properties:
        name:
          type: string
          description: Human-readable name (lowercase letters, digits, and dashes only; cannot start or end with a dash)
          pattern: ^[a-z0-9]([a-z0-9-]*[a-z0-9])?$
          maxLength: 63
          example: my-workload-1
        image:
          type: string
          description: OCI image reference
          example: docker.io/library/alpine:latest
        size:
          type: string
          description: Base memory size (human-readable format like "1GB", "512MB", "2G")
          default: "1GB"
          example: "2GB"
        hotplug_size:
          type: string
          description: Additional memory for hotplug (human-readable format like "3GB", "1G")
          default: "3GB"
          example: "2GB"
        overlay_size:
          type: string
          description: Writable overlay disk size (human-readable format like "10GB", "50G")
          default: "10GB"
          example: "20GB"
        disk_io_bps:
          type: string
          description: Disk I/O rate limit (e.g., "100MB/s", "500MB/s"). Defaults to proportional share based on CPU allocation if configured.
          example: "100MB/s"
        vcpus:
          type: integer
          description: Number of virtual CPUs
          default: 2
          example: 2
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
          example:
            PORT: "3000"
            NODE_ENV: production
        network:
          type: object
          description: Network configuration for the instance
          properties:
            enabled:
              type: boolean
              description: Whether to attach instance to the default network
              default: true
              example: true
            bandwidth_download:
              type: string
              description: Download bandwidth limit (external→VM, e.g., "1Gbps", "125MB/s"). Defaults to proportional share based on CPU allocation.
              example: "1Gbps"
            bandwidth_upload:
              type: string
              description: Upload bandwidth limit (VM→external, e.g., "1Gbps", "125MB/s"). Defaults to proportional share based on CPU allocation.
              example: "1Gbps"
        devices:
          type: array
          items:
            type: string
          description: Device IDs or names to attach for GPU/PCI passthrough
          example: ["l4-gpu"]
        volumes:
          type: array
          description: Volumes to attach to the instance at creation time
          items:
            $ref: "#/components/schemas/VolumeMount"
        hypervisor:
          type: string
          enum: [cloud-hypervisor, qemu]
          description: Hypervisor to use for this instance. Defaults to server configuration.
          example: cloud-hypervisor
        # Future: port_mappings, timeout_seconds
    Instance:
      type: object
      required: [id, name, image, state, created_at]
      properties:
        id:
          type: string
          description: Auto-generated unique identifier (CUID2 format)
          example: tz4a98xxat96iws9zmbrgj3a
        name:
          type: string
          description: Human-readable name
          example: my-workload-1
        image:
          type: string
          description: OCI image reference
          example: docker.io/library/alpine:latest
        state:
          $ref: "#/components/schemas/InstanceState"
        state_error:
          type: string
          description: Error message if state couldn't be determined (only set when state is Unknown)
          nullable: true
          example: "failed to query VMM: connection refused"
        size:
          type: string
          description: Base memory size (human-readable)
          example: "2GB"
        hotplug_size:
          type: string
          description: Hotplug memory size (human-readable)
          example: "2GB"
        overlay_size:
          type: string
          description: Writable overlay disk size (human-readable)
          example: "10GB"
        vcpus:
          type: integer
          description: Number of virtual CPUs
          example: 2
        disk_io_bps:
          type: string
          description: Disk I/O rate limit (human-readable, e.g., "100MB/s")
          example: "100MB/s"
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
        network:
          type: object
          description: Network configuration of the instance
          properties:
            enabled:
              type: boolean
              description: Whether instance is attached to the default network
              example: true
            name:
              type: string
              description: Network name (always "default" when enabled)
              example: "default"
            ip:
              type: string
              description: Assigned IP address (null if no network)
              example: "192.168.100.10"
              nullable: true
            mac:
              type: string
              description: Assigned MAC address (null if no network)
              example: "02:00:00:ab:cd:ef"
              nullable: true
            bandwidth_download:
              type: string
              description: Download bandwidth limit (human-readable, e.g., "1Gbps", "125MB/s")
              example: "125MB/s"
            bandwidth_upload:
              type: string
              description: Upload bandwidth limit (human-readable, e.g., "1Gbps", "125MB/s")
              example: "125MB/s"
        volumes:
          type: array
          description: Volumes attached to the instance
          items:
            $ref: "#/components/schemas/VolumeMount"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (RFC3339)
          example: "2025-01-15T10:30:00Z"
        started_at:
          type: string
          format: date-time
          description: Start timestamp (RFC3339)
          example: "2025-01-15T10:30:05Z"
          nullable: true
        stopped_at:
          type: string
          format: date-time
          description: Stop timestamp (RFC3339)
          example: "2025-01-15T12:30:00Z"
          nullable: true
        has_snapshot:
          type: boolean
          description: Whether a snapshot exists for this instance
          example: false
        hypervisor:
          type: string
          enum: [cloud-hypervisor, qemu]
          description: Hypervisor running this instance
          example: cloud-hypervisor
...
paths:
...
  /instances:
    get:
      summary: List instances
      operationId: listInstances
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instance"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create and start instance
      operationId: createInstance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstanceRequest"
      responses:
        201:
          description: Instance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...